Remove route-action on challenge
