+++ mon-pix/app/components/challenge/item.hbs
@@ -18,8 +18,8 @@
     resetAllChallengeInfo=@resetAllChallengeInfo
     resetChallengeInfoOnResume=@resetChallengeInfoOnResume
     hasFocusedOutOfWindow=@hasFocusedOutOfWindow
-    answerValidated=@answerValidated
-    resumeAssessment=@resumeAssessment
+    answerValidated=this.answerValidated
+    resumeAssessment=this.resumeAssessment
     isFocusedChallenge=this.isFocusedChallenge
   }}
 </article>
