+++ mon-pix/app/components/challenge/item.js
@@ -8,6 +8,8 @@ const FOCUSEDOUT_INTERVAL = 1000;
 
 export default class Item extends Component {
   @service currentUser;
+  @service router;
+  @service store;
 
   constructor() {
     super(...arguments);
@@ -68,5 +70,61 @@ export default class Item extends Component {
 
   get isFocusedChallenge() {
     return ENV.APP.FT_FOCUS_CHALLENGE_ENABLED && this.args.challenge.focused;
+  }
+
+  _findOrCreateAnswer(challenge, assessment) {
+    let answer = assessment.get('answers').findBy('challenge.id', challenge.id);
+    if (!answer) {
+      answer = this.store.createRecord('answer', { assessment, challenge });
+    }
+    return answer;
+  }
+
+  _isAssessmentEndedBySupervisorOrByFinalization(error) {
+    return (
+      error?.errors?.[0]?.detail === 'Le surveillant a mis fin à votre test de certification.' ||
+      error?.errors?.[0]?.detail === 'La session a été finalisée par votre centre de certification.'
+    );
+  }
+
+  @action
+  async answerValidated(challenge, assessment, answerValue, answerTimeout, answerFocusedOut) {
+    const answer = this._findOrCreateAnswer(challenge, assessment);
+    answer.setProperties({
+      value: answerValue.trim(),
+      timeout: answerTimeout,
+      focusedOut: answerFocusedOut,
+    });
+
+    try {
+      await answer.save();
+
+      let queryParams = { queryParams: {} };
+      const levelup = await answer.get('levelup');
+
+      if (this.currentUser.user && !this.currentUser.user.isAnonymous && levelup) {
+        queryParams = {
+          queryParams: {
+            newLevel: levelup.level,
+            competenceLeveled: levelup.competenceName,
+          },
+        };
+      }
+
+      this.router.transitionTo('assessments.resume', assessment.get('id'), queryParams);
+    } catch (error) {
+      answer.rollbackAttributes();
+
+      if (this._isAssessmentEndedBySupervisorOrByFinalization(error)) {
+        this.router.transitionTo('authenticated.certifications.results', assessment.certificationCourse.get('id'));
+      }
+
+      this.router.transitionTo('error', error);
+    }
+  }
+
+  @action
+  resumeAssessment(assessment) {
+    this.router.transitionTo('assessments.resume', assessment.get('id'));
   }
 }
