+++ mon-pix/app/templates/assessments/challenge.hbs
@@ -62,8 +62,6 @@
         @timeoutChallenge={{this.timeoutChallenge}}
         @resetAllChallengeInfo={{this.resetAllChallengeInfo}}
         @resetChallengeInfoOnResume={{this.resetChallengeInfoOnResume}}
-        @answerValidated={{route-action "saveAnswerAndNavigate"}}
-        @resumeAssessment={{route-action "resumeAssessment"}}
         @onFocusIntoChallenge={{fn this.setFocusedOutOfChallenge false}}
         @onFocusOutOfChallenge={{fn this.setFocusedOutOfChallenge true}}
         @onFocusOutOfWindow={{this.focusedOutOfWindow}}
