+++ mon-pix/tests/unit/routes/assessments/challenge_test.js
@@ -1,6 +1,5 @@
 import { describe, it } from 'mocha';
 import { setupTest } from 'ember-mocha';
-import EmberObject from '@ember/object';
 import EmberService from '@ember/service';
 import sinon from 'sinon';
 
@@ -143,284 +142,6 @@ describe('Unit | Route | Assessments | Challenge', function () {
 
         // then
         sinon.assert.calledWith(findRecordStub, 'challenge', 'oldRecId');
-      });
-    });
-  });
-
-  describe('#saveAnswerAndNavigate', function () {
-    let answerToChallengeOne, answerToChallengeOneWithLevelUp;
-
-    let answerValue = 'example';
-    const answerTimeout = 120;
-    const answerFocusedOut = false;
-    const challengeOne = EmberObject.create({ id: 'recChallengeOne' });
-    const nextChallenge = EmberObject.create({ id: 'recNextChallenge' });
-
-    beforeEach(function () {
-      answerToChallengeOne = EmberObject.create({ challenge: challengeOne });
-      answerToChallengeOne.save = sinon.stub().resolves();
-      answerToChallengeOne.setProperties = sinon.stub();
-      answerToChallengeOne.rollbackAttributes = sinon.stub();
-
-      answerToChallengeOneWithLevelUp = EmberObject.create({
-        challenge: challengeOne,
-        levelup: { level: 1, competenceName: 'Me tester' },
-      });
-      answerToChallengeOneWithLevelUp.save = sinon.stub().resolves();
-      answerToChallengeOneWithLevelUp.setProperties = sinon.stub();
-      answerToChallengeOneWithLevelUp.rollbackAttributes = sinon.stub();
-    });
-
-    context('when the answer is already known', function () {
-      it('should not create a new answer', function () {
-        // given
-        const assessment = EmberObject.create({ answers: [answerToChallengeOne] });
-        createRecordStub.returns(answerToChallengeOne);
-        queryRecordStub.resolves(nextChallenge);
-
-        // when
-        route.actions.saveAnswerAndNavigate.call(route, challengeOne, assessment, answerValue, answerTimeout);
-
-        // then
-        sinon.assert.notCalled(createRecordStub);
-      });
-    });
-
-    context('when no answer was given', function () {
-      it('should create an answer', function () {
-        // given
-        const assessment = EmberObject.create({ answers: [] });
-        createRecordStub.returns(answerToChallengeOne);
-        queryRecordStub.resolves(nextChallenge);
-
-        // when
-        route.actions.saveAnswerAndNavigate.call(route, challengeOne, assessment, answerValue, answerTimeout);
-
-        // then
-        sinon.assert.calledWith(createRecordStub, 'answer', {
-          assessment: assessment,
-          challenge: challengeOne,
-        });
-      });
-    });
-
-    it('should update the answer with the timeout', function () {
-      // given
-      const assessment = EmberObject.create({ answers: [answerToChallengeOne] });
-      createRecordStub.returns(answerToChallengeOne);
-      queryRecordStub.resolves(nextChallenge);
-
-      // when
-      route.actions.saveAnswerAndNavigate.call(
-        route,
-        challengeOne,
-        assessment,
-        answerValue,
-        answerTimeout,
-        answerFocusedOut
-      );
-
-      // then
-      sinon.assert.callOrder(answerToChallengeOne.setProperties, answerToChallengeOne.save);
-      sinon.assert.calledOnce(answerToChallengeOne.save);
-      sinon.assert.calledWith(answerToChallengeOne.setProperties, {
-        value: answerValue,
-        timeout: answerTimeout,
-        focusedOut: answerFocusedOut,
-      });
-    });
-
-    it('should trim the answer value to avoid useless char', function () {
-      // given
-      answerValue = '  exemple \n ';
-      const answerValueWithoutUselessChar = 'exemple';
-      const assessment = EmberObject.create({ answers: [answerToChallengeOne] });
-      createRecordStub.returns(answerToChallengeOne);
-      queryRecordStub.resolves(nextChallenge);
-
-      // when
-      route.actions.saveAnswerAndNavigate.call(
-        route,
-        challengeOne,
-        assessment,
-        answerValue,
-        answerTimeout,
-        answerFocusedOut
-      );
-
-      // then
-      sinon.assert.callOrder(answerToChallengeOne.setProperties, answerToChallengeOne.save);
-      sinon.assert.calledOnce(answerToChallengeOne.save);
-      sinon.assert.calledWith(answerToChallengeOne.setProperties, {
-        value: answerValueWithoutUselessChar,
-        timeout: answerTimeout,
-        focusedOut: answerFocusedOut,
-      });
-    });
-
-    context('when saving succeeds', function () {
-      it('should redirect to assessment-resume route', async function () {
-        // given
-        currentUserStub.isAnonymous = false;
-        route.currentUser = currentUserStub;
-        const assessment = EmberObject.create({ answers: [answerToChallengeOne] });
-
-        // when
-        await route.actions.saveAnswerAndNavigate.call(
-          route,
-          challengeOne,
-          assessment,
-          answerValue,
-          answerTimeout,
-          answerFocusedOut
-        );
-
-        // then
-        sinon.assert.calledWithExactly(route.router.transitionTo, 'assessments.resume', assessment.get('id'), {
-          queryParams: {},
-        });
-      });
-
-      context('when user has reached a new level', function () {
-        let assessment;
-        beforeEach(function () {
-          createRecordStub.returns(answerToChallengeOneWithLevelUp);
-          queryRecordStub.resolves(nextChallenge);
-          assessment = EmberObject.create({ answers: [answerToChallengeOneWithLevelUp] });
-        });
-
-        it('should redirect to assessment-resume route with level up information', async function () {
-          //given
-          currentUserStub.user.isAnonymous = false;
-          route.currentUser = currentUserStub;
-          const expectedQueryParams = {
-            queryParams: {
-              newLevel: answerToChallengeOneWithLevelUp.levelup.level,
-              competenceLeveled: answerToChallengeOneWithLevelUp.levelup.competenceName,
-            },
-          };
-
-          // when
-          await route.actions.saveAnswerAndNavigate.call(
-            route,
-            challengeOne,
-            assessment,
-            answerValue,
-            answerTimeout,
-            answerFocusedOut
-          );
-
-          // then
-          sinon.assert.calledWithExactly(
-            route.router.transitionTo,
-            'assessments.resume',
-            assessment.get('id'),
-            expectedQueryParams
-          );
-        });
-
-        it('should redirect to assessment-resume route without level up information when user is anonymous', async function () {
-          //given
-          currentUserStub.user.isAnonymous = true;
-          route.currentUser = currentUserStub;
-          const expectedQueryParams = { queryParams: {} };
-
-          // when
-          await route.actions.saveAnswerAndNavigate.call(
-            route,
-            challengeOne,
-            assessment,
-            answerValue,
-            answerTimeout,
-            answerFocusedOut
-          );
-
-          // then
-          sinon.assert.calledWithExactly(
-            route.router.transitionTo,
-            'assessments.resume',
-            assessment.get('id'),
-            expectedQueryParams
-          );
-        });
-
-        it('should redirect to assessment-resume route without level up information when there is no currentUser', async function () {
-          //given
-          route.currentUser = { user: undefined };
-          const expectedQueryParams = { queryParams: {} };
-
-          // when
-          await route.actions.saveAnswerAndNavigate.call(
-            route,
-            challengeOne,
-            assessment,
-            answerValue,
-            answerTimeout,
-            answerFocusedOut
-          );
-
-          // then
-          sinon.assert.calledWithExactly(
-            route.router.transitionTo,
-            'assessments.resume',
-            assessment.get('id'),
-            expectedQueryParams
-          );
-        });
-      });
-    });
-
-    context('when saving fails', function () {
-      it('should remove temporary answer and send error', async function () {
-        // given
-        const error = { message: 'error' };
-        answerToChallengeOne.save = sinon.stub().rejects(error);
-        route.intermediateTransitionTo = sinon.stub().rejects();
-        const assessment = EmberObject.create({ answers: [answerToChallengeOne] });
-
-        // when / then
-        return route.actions.saveAnswerAndNavigate
-          .call(route, challengeOne, assessment, answerValue, answerTimeout)
-          .then(function () {
-            throw new Error('was supposed to fail');
-          })
-          .catch(function () {
-            sinon.assert.called(answerToChallengeOne.rollbackAttributes);
-            sinon.assert.calledWith(route.intermediateTransitionTo, 'error', error);
-          });
-      });
-    });
-
-    context('when assessmnt has been ended by supervisor', function () {
-      it('should redirect candidate to end test screen when trying to answer', async function () {
-        // given
-        const error = {
-          errors: [
-            {
-              detail: 'Le surveillant a mis fin à votre test de certification.',
-            },
-          ],
-        };
-        answerToChallengeOne.save = sinon.stub().rejects(error);
-        const certificationCourse = EmberObject.create({});
-        const assessment = EmberObject.create({ answers: [answerToChallengeOne], certificationCourse });
-
-        // when
-        await route.actions.saveAnswerAndNavigate.call(
-          route,
-          challengeOne,
-          assessment,
-          answerValue,
-          answerTimeout,
-          answerFocusedOut
-        );
-
-        // then
-        sinon.assert.calledWithExactly(
-          route.router.transitionTo,
-          'authenticated.certifications.results',
-          assessment.certificationCourse.get('id')
-        );
       });
     });
   });
